split off net gradient
