+++ src/caffe2/util/net.cc
@@ -9,54 +9,6 @@
 
 namespace caffe2 {
 
-const std::set<std::string> trainable_ops({
-    "Add",
-    "AffineScale",
-    "AveragedLoss",
-    "AveragePool",
-    "BackMean",
-    "Concat",
-    "Conv",
-    "Diagonal",
-    "Dropout",
-    "EnsureCPUOutput",
-    "FC",
-    "LabelCrossEntropy",
-    "LRN",
-    "MaxPool",
-    "Mul",
-    "RecurrentNetwork",
-    "Relu",
-    "Reshape",
-    "Slice",
-    "Softmax",
-    "SpatialBN",
-    "SquaredL2",
-    "SquaredL2Channel",
-    "Sum",
-});
-
-const std::set<std::string> non_trainable_ops({
-    "Accuracy", "Cast", "Cout", "ConstantFill", "Iter", "Scale", "StopGradient",
-    "TensorProtosDBInput", "TimePlot", "ShowWorst",
-});
-
-const std::map<std::string, std::string> custom_gradient({
-    {"EnsureCPUOutput", "CopyFromCPUInput"},
-    {"CopyFromCPUInput", "EnsureCPUOutput"},
-});
-
-const std::set<std::string> filler_ops({
-    "UniformFill", "UniformIntFill", "UniqueUniformFill", "ConstantFill",
-    "GaussianFill", "XavierFill", "MSRAFill", "RangeFill", "LengthsRangeFill",
-});
-
-const std::set<std::string> non_inplace_ops({
-    "Dropout",  // TODO: see if "they" fixed dropout on cudnn
-});
-
-const std::string gradient_suffix("_grad");
-
 // Helpers
 
 OperatorDef* NetUtil::AddOp(const std::string& name,
@@ -686,6 +638,10 @@ void NetUtil::SetFillToTrain() {
   }
 }
 
+const std::set<std::string> non_inplace_ops({
+    "Dropout",  // TODO: see if "they" fixed dropout on cudnn
+});
+
 void NetUtil::SetRenameInplace() {
   std::set<std::string> renames;
   for (auto& op : *net.mutable_op()) {
@@ -707,177 +663,6 @@ void NetUtil::SetRenameInplace() {
 void NetUtil::SetEngineOps(const std::string engine) {
   for (auto& op : *net.mutable_op()) {
     op.set_engine(engine);
-  }
-}
-
-// Gradient
-
-OperatorDef* NetUtil::AddGradientOp(
-    OperatorDef& op, std::map<std::string, std::pair<int, int>>& split_inputs) {
-  auto grad = net.add_op();
-  if (custom_gradient.find(op.type()) == custom_gradient.end()) {
-    vector<GradientWrapper> output(op.output_size());
-    for (auto i = 0; i < output.size(); i++) {
-      output[i].dense_ = op.output(i) + gradient_suffix;
-    }
-    GradientOpsMeta meta = GetGradientForOp(op, output);
-    if (meta.ops_.size()) {
-      if (meta.ops_.size() > 1) {
-        std::cerr << "multiple gradients for operator (" << op.type();
-        for (auto& o : meta.ops_) {
-          std::cerr << " " << o.type();
-        }
-        std::cerr << ")" << std::endl;
-      }
-      grad->CopyFrom(meta.ops_[0]);
-    } else {
-      std::cerr << "no gradient for operator " << op.type() << std::endl;
-    }
-  } else {
-    grad->set_type(custom_gradient.at(op.type()));
-    for (auto arg : op.arg()) {
-      auto copy = grad->add_arg();
-      copy->CopyFrom(arg);
-    }
-    for (auto output : op.output()) {
-      grad->add_input(output + gradient_suffix);
-    }
-    for (auto input : op.input()) {
-      grad->add_output(input + gradient_suffix);
-    }
-  }
-  grad->set_is_gradient_op(true);
-  for (auto i = 0; i < grad->output_size(); i++) {
-    auto output = grad->output(i);
-    if (split_inputs.count(output)) {
-      split_inputs[output].first--;
-      grad->set_output(
-          i, output + "_sum_" + std::to_string(split_inputs[output].first));
-      if (split_inputs[output].first == 0) {
-        std::vector<std::string> inputs;
-        for (int i = 0; i < split_inputs[output].second; i++) {
-          inputs.push_back(output + "_sum_" + std::to_string(i));
-        }
-        AddSumOp(inputs, output);
-      }
-    }
-  }
-  return grad;
-}
-
-void NetUtil::AddAllGradientOp() {
-  std::map<std::string, std::pair<int, int>> split_inputs;
-  for (auto op : CollectGradientOps(split_inputs)) {
-    AddGradientOp(op, split_inputs);
-  }
-}
-
-// Collectors
-
-std::map<std::string, int> NetUtil::CollectParamSizes() {
-  std::map<std::string, int> sizes;
-  for (const auto& op : net.op()) {
-    if (filler_ops.find(op.type()) != filler_ops.end()) {
-      for (const auto& arg : op.arg()) {
-        if (arg.name() == "shape") {
-          auto size = 1;
-          for (auto i : arg.ints()) {
-            size *= i;
-          }
-          sizes[op.output(0)] = size;
-        }
-      }
-    }
-  }
-  return sizes;
-}
-
-std::vector<std::string> NetUtil::CollectParams() {
-  std::vector<std::string> params;
-  std::set<std::string> external_inputs(net.external_input().begin(),
-                                        net.external_input().end());
-  for (const auto& op : net.op()) {
-    if (trainable_ops.find(op.type()) != trainable_ops.end()) {
-      for (const auto& input : op.input()) {
-        if (external_inputs.find(input) != external_inputs.end()) {
-          params.push_back(input);
-        }
-      }
-    }
-  }
-  return params;
-}
-
-std::vector<OperatorDef> NetUtil::CollectGradientOps(
-    std::map<std::string, std::pair<int, int>>& split_inputs) {
-  std::set<std::string> external_inputs(net.external_input().begin(),
-                                        net.external_input().end());
-  std::vector<OperatorDef> gradient_ops;
-  std::map<std::string, int> input_count;
-  for (auto& op : net.op()) {
-    if (trainable_ops.find(op.type()) != trainable_ops.end()) {
-      gradient_ops.push_back(op);
-      for (auto& input : op.input()) {
-        auto& output = op.output();
-        if (std::find(output.begin(), output.end(), input) == output.end()) {
-          input_count[input]++;
-          if (input_count[input] > 1) {
-            split_inputs[input + "_grad"] = {input_count[input],
-                                             input_count[input]};
-          }
-        }
-      }
-    } else if (non_trainable_ops.find(op.type()) == non_trainable_ops.end()) {
-      CAFFE_THROW("unknown backprop operator type: " + op.type());
-    }
-  }
-  std::reverse(gradient_ops.begin(), gradient_ops.end());
-  return gradient_ops;
-}
-
-std::set<std::string> NetUtil::CollectLayers(const std::string& layer,
-                                             bool forward) {
-  std::map<std::string, std::set<std::string>> lookup;
-  for (auto& op : net.op()) {
-    for (auto& input : op.input()) {
-      for (auto& output : op.output()) {
-        lookup[forward ? input : output].insert(forward ? output : input);
-      }
-    }
-  }
-  std::set<std::string> result;
-  for (std::set<std::string> step({layer}); step.size();) {
-    std::set<std::string> next;
-    for (auto& l : step) {
-      if (result.find(l) == result.end()) {
-        result.insert(l);
-        for (auto& n : lookup[l]) {
-          next.insert(n);
-        }
-      }
-    }
-    step = next;
-  }
-  return result;
-}
-
-void NetUtil::CheckLayerAvailable(const std::string& layer) {
-  std::vector<std::pair<std::string, std::string>> available_layers(
-      {{net.external_input(0), "Input"}});
-  auto layer_found = (net.external_input(0) == layer);
-  for (const auto& op : net.op()) {
-    if (op.input(0) != op.output(0)) {
-      available_layers.push_back({op.output(0), op.type()});
-      layer_found |= (op.output(0) == layer);
-    }
-  }
-  if (!layer_found) {
-    std::cerr << "available layers:" << std::endl;
-    for (auto& layer : available_layers) {
-      std::cerr << "  " << layer.first << " (" << layer.second << ")"
-                << std::endl;
-    }
-    LOG(FATAL) << "~ no layer with name " << layer << " in model.";
   }
 }
